+++ src/modules/transactions/actions/trigger-transactions-export.test.js
@@ -2,6 +2,7 @@ import configureMockStore from "redux-mock-store";
 import thunk from "redux-thunk";
 import { triggerTransactionsExport } from "modules/transactions/actions/trigger-transactions-export";
 import * as mockLoadAccountHistory from "modules/auth/actions/load-account-history";
+import { selectTransactions } from "src/modules/transactions/selectors/transactions";
 
 jest.mock("modules/transactions/selectors/transactions");
 jest.mock("modules/auth/actions/load-account-history");
@@ -9,10 +10,7 @@ jest.mock("modules/auth/actions/load-account-history");
 describe("modules/transactions/actions/trigger-transactions-export.js", () => {
   const middlewares = [thunk];
   const mockStore = configureMockStore(middlewares);
-
-  beforeEach(() => {
-    jest.resetModules();
-  });
+  selectTransactions.mockImplementation(state => state.transactions);
 
   test("if triggers a download if transactionsLoading is false", () => {
     const store = mockStore({
