--- src/modules/transactions/selectors/__mocks__/transactions.js
@@ -1,7 +1,0 @@
-const mockTransactions = jest.genMockFromModule(
-  "modules/transactions/selectors/transactions.js"
-);
-
-mockTransactions.selectTransactions = state => state.transactions;
-
-module.exports = mockTransactions;
